+++ src/App.js
@@ -1,9 +1,15 @@
 import React, { Component } from 'react'
 import Console from './console/Console'
+import HistoryListContainer from './history/containers/HistoryListContainer'
 
 class App extends Component {
   render() {
-    return <Console />
+    return (
+      <div>
+        <Console />
+        <HistoryListContainer />
+      </div>
+    )
   }
 }
