+++ src/console/Console.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react'
 import Form from './Form'
 import Display from './Display'
 import makeApiRequest from './api-tester'
-import HistoryListContainer from './containers/HistoryListContainer'
 
 class Console extends Component {
   constructor(props) {
@@ -23,7 +22,6 @@ class Console extends Component {
       <div>
         <Display response={this.state.response} />
         <Form submitFormData={this.testApi} />
-        <HistoryListContainer/>
       </div>
     )
   }
