--- src/console/containers/HistoryListContainer.js
@@ -1,12 +1,0 @@
-import { connect } from 'react-redux'
-import HistoryList from '../HistoryList'
-
-const mapStateToProps = (state, ownProps) => ({ 
-    historyList : state
-});
-
-const HistoryListContainer = connect(  
-    mapStateToProps
-)(HistoryList)
-
-export default HistoryListContainer
