--- src/console/store.js
@@ -1,9 +1,0 @@
-import reducers from './reducers/HistoryReducer'
-import {createStore} from 'redux';
-
-export function configureStore(initialState = []) {  
-    const store = createStore(reducers, initialState)
-    return store
-  };
-  
-export const store = configureStore();
