+++ src/index.js
@@ -2,11 +2,12 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
-import { store } from './console/store'
+import { store } from './store/store'
 import { Provider } from 'react-redux'
 
 ReactDOM.render(
-    <Provider store={store}>
-        <App />
-    </Provider>, 
-    document.getElementById('root'))
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  document.getElementById('root')
+)
